Tests: libraries util
